change ini logic, add support for opening from strings

+++ OpenGM/VirtualMachine/BuiltInFunctions/FileFunctions.cs
@@ -307,7 +307,20 @@ public static class FileFunctions
         return Environment.GetEnvironmentVariable(name);
     }
 
-    // ini_open_from_string
+    [GMLFunction("ini_open_from_string")]
+    public static object? ini_open_from_string(object?[] args)
+    {
+        var content = args[0].Conv<string>();
+
+        if (_iniFile != null)
+        {
+            ini_close(new object[0]);
+        }
+
+        _iniFile = IniFile.FromContent(content);
+
+        return null;
+    }
 
     [GMLFunction("ini_open")]
     public static object? ini_open(object?[] args)
@@ -331,38 +344,9 @@ public static class FileFunctions
             return null;
         }
 
-        var lines = File.ReadAllLines(filepath);
-
-        KeyValuePair<string, string> ParseKeyValue(string line)
-        {
-            var lineByEquals = line.Split('=');
-            var key = lineByEquals[0].Trim();
-            var value = lineByEquals[1].Trim();
-            value = value.Trim('"');
-            return new KeyValuePair<string, string>(key, value);
-        }
-
-        _iniFile = new IniFile { Name = name };
-        IniSection? currentSection = null;
-
-        for (var i = 0; i < lines.Length; i++)
-        {
-            var currentLine = lines[i];
-            if (currentLine.StartsWith('[') && currentLine.EndsWith("]"))
-            {
-                currentSection = new IniSection(currentLine.TrimStart('[').TrimEnd(']'));
-                _iniFile.Sections.Add(currentSection);
-                continue;
-            }
-
-            if (string.IsNullOrEmpty(currentLine))
-            {
-                continue;
-            }
-
-            var keyvalue = ParseKeyValue(currentLine);
-            currentSection?.Dict.Add(keyvalue.Key, keyvalue.Value);
-        }
+        var text = File.ReadAllText(filepath);
+        _iniFile = IniFile.FromContent(text);
+        _iniFile.Name = name;
 
         return null;
     }
@@ -375,27 +359,23 @@ public static class FileFunctions
             return null;
         }
 
-        var filepath = Path.Combine(Entry.DataWinFolder, _iniFile!.Name);
-        File.Delete(filepath);
-        var fileStream = new FileStream(filepath, FileMode.Append, FileAccess.Write);
-        var streamWriter = new StreamWriter(fileStream);
+        var text = _iniFile.ToString();
 
-        foreach (var section in _iniFile.Sections)
+        if (_iniFile.Name is not null)
         {
-            streamWriter.WriteLine($"[{section.Name}]");
-            foreach (var kv in section.Dict)
-            {
-                streamWriter.WriteLine($"{kv.Key}=\"{kv.Value}\"");
-            }
+            var filepath = Path.Combine(Entry.DataWinFolder, _iniFile!.Name);
+            File.Delete(filepath);
+            var fileStream = new FileStream(filepath, FileMode.Append, FileAccess.Write);
+            var streamWriter = new StreamWriter(fileStream);
+
+            streamWriter.Write(text);
+            streamWriter.Close();
+            streamWriter.Dispose();
         }
 
-        var text = streamWriter.ToString();
-
-        streamWriter.Close();
-        streamWriter.Dispose();
         _iniFile = null;
 
-        return text; // BUG: this does NOT return the written text
+        return text;
     }
 
     [GMLFunction("ini_read_string")]
